requirements.txt
# LLM Council - Python Dependencies

# LLM API
openai>=1.0.0

# Data validation
pydantic>=2.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# UI
gradio>=4.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
